ole/requirements.txt
openai==1.3.2
pillow
pybase64